# openapi.yaml
openapi: 3.1.0
info:
  title: AgriCRM API
  description: Farmer, FPO & Sugar Mill CRM for Theta Analytics
  version: 1.0.0
paths:
...
  /api/v1/people/:
    get:
      tags:
      - people
      summary: Person List
      description: "The people register. **Names and posts only \u2014 no contact\
        \ values.**\n\nFilters are declared parameters and `StrictQuery` rejects anything\
        \ else\nwith a 400: a typo'd filter that silently does nothing is how someone\n\
        exports the whole register believing they exported one district."
      operationId: person_list_api_v1_people__get
      security:
      - HTTPBearer: []
      parameters:
      - name: organisation
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Organisation
      - name: role
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: State
      - name: district
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: District
      - name: quality_tier
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Quality Tier
      - name: is_farmer
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Farmer
      - name: din
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Din
      - name: current_roles_only
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Current Roles Only
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Deleted
      - name: reason
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Required (R10) when the result exceeds the bulk threshold.
            Recorded against your account in audit.data_access_log.
          title: Reason
        description: Required (R10) when the result exceeds the bulk threshold. Recorded
          against your account in audit.data_access_log.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - people
      summary: Person Create
      description: "Record a person.\n\n\U0001F534 The `source_id` gate is the whole\
        \ point of this endpoint. R4 permits\npersonal data through four routes, and\
        \ `require_pii_source` is where that\nsentence becomes something a request\
        \ either passes or fails."
      operationId: person_create_api_v1_people__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/people/{person_id}:
    get:
      tags:
      - people
      summary: Person Detail
      description: "One person, with every post they have held and every way to reach\
        \ them.\n\n\U0001F534 `unmask=true` is checked *before* the response is assembled\
        \ and is\nlogged whether or not the person turns out to have any contact points.\
        \ A\ncaller who asks to see personal data has asked, and that is the fact\
        \ the\naudit log records."
      operationId: person_detail_api_v1_people__person_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: person_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Person Id
      - name: unmask
        in: query
        required: false
        schema:
          type: boolean
          description: Show full contact values. Requires the contact.view_full capability
            (R9); the view is written to audit.data_access_log.
          default: false
          title: Unmask
        description: Show full contact values. Requires the contact.view_full capability
          (R9); the view is written to audit.data_access_log.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - people
      summary: Person Update
      description: Correct a person's details. Provenance and tier are not editable
        here.
      operationId: person_update_api_v1_people__person_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: person_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Person Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/people/{person_id}/roles:
    post:
      tags:
      - people
      summary: Person Role Create
      description: "Record that this person holds a post at an organisation.\n\n\U0001F534\
        \ An open role is not replaced by a new one \u2014 both are written, and the\n\
        old one is closed through `PATCH .../roles/{id}` with a date. That is what\n\
        lets the register answer \"who was chairman in March 2025\" afterwards."
      operationId: person_role_create_api_v1_people__person_id__roles_post
      security:
      - HTTPBearer: []
      parameters:
      - name: person_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Person Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/people/{person_id}/roles/{role_id}:
    patch:
      tags:
      - people
      summary: Person Role Close
      description: 'End a post by dating it. There is no delete.


        The DDL''s CHECK requires `valid_to >= valid_from`; a date before the start

        is rejected there rather than re-implemented here.'
      operationId: person_role_close_api_v1_people__person_id__roles__role_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: person_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Person Id
      - name: role_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Role Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCloseIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/people/by-organisation/{organisation_id}:
    get:
      tags:
      - people
      summary: Organisation People
      description: "Who holds which post at this organisation \u2014 the board, in\
        \ other words.\n\nNo contact values. This is the endpoint a BD screen calls\
        \ to show \"MD:\nSunita Devi\", and it is deliberately not the endpoint that\
        \ would hand a\nclient every director's mobile in one response."
      operationId: organisation_people_api_v1_people_by_organisation__organisation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: organisation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Organisation Id
      - name: current_only
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Current Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleOut'
                title: Response Organisation People Api V1 People By Organisation  Organisation
                  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PersonIn:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        middle_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Middle Name
        salutation:
          anyOf:
          - type: string
          - type: 'null'
          title: Salutation
        name_local:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Local
        father_or_spouse:
          anyOf:
          - type: string
          - type: 'null'
          title: Father Or Spouse
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
        date_of_birth:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Date Of Birth
        din:
          anyOf:
          - type: string
          - type: 'null'
          title: Din
        state_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: State Id
        district_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: District Id
        village_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Village Id
        is_farmer:
          type: boolean
          title: Is Farmer
          default: false
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        source_id:
          type: integer
          title: Source Id
      type: object
      required:
      - first_name
      - source_id
      title: PersonIn
      description: "A person being created.\n\n\U0001F534 `source_id` is required\
        \ for the same reason it is on a contact point,\nand it is checked against\
        \ `dq.source` before the row is written: R4 says\npersonal data enters only\
        \ through four named routes, and a source row is\nhow the system knows which\
        \ one this was."
    PersonPage:
      properties:
        count:
          type: integer
          title: Count
        results:
          items:
            $ref: '#/components/schemas/PersonRow'
          type: array
          title: Results
      type: object
      required:
      - count
      - results
      title: PersonPage
...
    RoleCloseIn:
      properties:
        valid_to:
          type: string
          format: date
          title: Valid To
      type: object
      required:
      - valid_to
      title: RoleCloseIn
      description: "\U0001F534 Closing a role, which is the only way a role ends.\n\
        \nThere is no delete endpoint for a role and this is the reason: the\nregister\
        \ has to answer who held a post on a past date, and a deleted row\nanswers\
        \ \"nobody, ever\"."
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer